+++ efcms-service/src/applicationContext.js
@@ -6,6 +6,9 @@ const {
 } = require('ef-cms-shared/src/persistence/dynamo/helpers/incrementCounter');
 
 const {
+  getCasesForRespondent
+} = require('ef-cms-shared/src/persistence/dynamo/cases/getCasesForRespondent');
+const {
   getSentWorkItemsForUser,
 } = require('ef-cms-shared/src/persistence/dynamo/workitems/getSentWorkItemsForUser');
 const {
@@ -41,9 +44,6 @@ const {
   getCasesByUser,
 } = require('ef-cms-shared/src/persistence/dynamo/cases/getCasesByUser');
 const {
-  getCasesForRespondent,
-} = require('ef-cms-shared/src/persistence/dynamo/cases/getCasesForRespondent');
-const {
   saveCase,
 } = require('ef-cms-shared/src/persistence/dynamo/cases/saveCase');
 const {
@@ -74,14 +74,8 @@ const {
   getCase,
 } = require('ef-cms-shared/src/business/useCases/getCase.interactor');
 const {
-  getCasesByStatus: getCasesByStatusUC,
-} = require('ef-cms-shared/src/business/useCases/getCasesByStatus.interactor');
-const {
   createCase,
 } = require('ef-cms-shared/src/business/useCases/createCase.interactor');
-const {
-  getCasesByUser: getCasesByUserUC,
-} = require('ef-cms-shared/src/business/useCases/getCasesByUser.interactor');
 const {
   getInternalUsers: getInternalUsersUC,
 } = require('ef-cms-shared/src/business/useCases/users/getUsersInSection.interactor')
@@ -95,8 +89,8 @@ const {
   updateCase,
 } = require('ef-cms-shared/src/business/useCases/updateCase.interactor');
 const {
-  getCasesForRespondent: getCasesForRespondentUC,
-} = require('ef-cms-shared/src/business/useCases/respondent/getCasesForRespondent.interactor');
+  queryForCases,
+} = require('ef-cms-shared/src/business/useCases/cases/queryForCases.interactor');
 const {
   getWorkItem,
 } = require('ef-cms-shared/src/business/useCases/workitems/getWorkItem.interactor');
@@ -109,9 +103,6 @@ const {
 const {
   createDocument,
 } = require('ef-cms-shared/src/business/useCases/createDocument.interactor');
-const {
-  getInteractorForGettingCases,
-} = require('ef-cms-shared/src/business/useCases/utilities/getInteractorForGettingCases');
 const {
   getInteractorForGetUsers,
 } = require('ef-cms-shared/src/business/useCases/utilities/getInteractorForGetUsers');
@@ -239,19 +230,17 @@ module.exports = (appContextUser = {}) => {
         createCase,
         createUser: createUserUC,
         getCase,
-        getCasesByStatus: getCasesByStatusUC,
-        getCasesByUser: getCasesByUserUC,
         getUsersInSection: getUsersInSectionUC,
         getInternalUsers: getInternalUsersUC,
         getUser,
         forwardWorkItem,
         sendPetitionToIRSHoldingQueue,
         updateCase,
-        getCasesForRespondent: getCasesForRespondentUC,
         getWorkItem,
         getWorkItems,
         getWorkItemsUC,
         updateWorkItem,
+        queryForCases,
         createDocument,
         getWorkItemsBySection: getWorkItemsBySectionUC,
         getSentWorkItemsForSection: getSentWorkItemsForSectionUC,
@@ -291,7 +280,6 @@ module.exports = (appContextUser = {}) => {
         return getWorkItemsUC;
       }
     },
-    getInteractorForGettingCases,
     getInteractorForGetUsers,
   };
 };
