+++ efcms-service/storage/fixtures/efcms.json
@@ -452,7 +452,7 @@
   },
   {
     "sk": "e631d81f-a579-4de5-b8a8-b3f10ef619fd",
-    "pk": "respondent|activeCase"
+    "pk": "respondent|case"
   },
   {
     "sk": "e631d81f-a579-4de5-b8a8-b3f10ef619fd",
