--- shared/src/business/useCases/getCasesByDocumentId.interactor.test.js
@@ -1,88 +1,0 @@
-const { getCasesByDocumentId } = require('./getCasesByDocumentId.interactor');
-
-describe('getCasesByDocumentId', () => {
-  let applicationContext;
-  it('throws an error when the user is not validated to get the cases by status', async () => {
-    applicationContext = {
-      getPersistenceGateway: () => {
-        return {
-          getCasesByDocumentId: async () => [
-            { caseId: 'c54ba5a9-b37b-479d-9201-067ec6e335bb' },
-          ],
-        };
-      },
-      getCurrentUser: () => {
-        return {
-          userId: 'taxpayer',
-          role: 'petitioner',
-        };
-      },
-      environment: { stage: 'local' },
-    };
-    let error;
-    try {
-      await getCasesByDocumentId({
-        documentId: 'c54ba5a9-b37b-479d-9201-067ec6e335bb',
-        applicationContext,
-      });
-    } catch (e) {
-      error = e;
-    }
-    expect(error).toBeDefined();
-  });
-
-  it('returns the cases returned from persistence', async () => {
-    applicationContext = {
-      getPersistenceGateway: () => {
-        return {
-          getCasesByDocumentId: async () => [
-            {
-              caseId: 'c54ba5a9-b37b-479d-9201-067ec6e335bb',
-              docketNumber: '101-18',
-              petitioners: [{ name: 'Test Taxpayer' }],
-              caseType: 'Other',
-              procedureType: 'Regular',
-              filingType: 'Myself',
-              preferredTrialCity: 'Washington, D.C.',
-              documents: [
-                {
-                  documentType: 'Answer',
-                  userId: 'respondent',
-                  documentId: 'c54ba5a9-b37b-479d-9201-067ec6e335bb',
-                },
-              ],
-            },
-          ],
-        };
-      },
-      getCurrentUser: () => {
-        return {
-          userId: 'docketclerk',
-          role: 'docketclerk',
-        };
-      },
-      environment: { stage: 'local' },
-    };
-    const result = await getCasesByDocumentId({
-      documentId: 'c54ba5a9-b37b-479d-9201-067ec6e335bb',
-      applicationContext,
-    });
-
-    expect(result).toMatchObject([
-      {
-        caseId: 'c54ba5a9-b37b-479d-9201-067ec6e335bb',
-        docketNumber: '101-18',
-        petitioners: [{ name: 'Test Taxpayer' }],
-        documents: [
-          {
-            documentId: 'c54ba5a9-b37b-479d-9201-067ec6e335bb',
-            documentType: 'Answer',
-            userId: 'respondent',
-            workItems: [],
-          },
-        ],
-        status: 'New',
-      },
-    ]);
-  });
-});
