--- shared/src/business/useCases/respondent/getCasesForRespondent.interactor.js
@@ -1,19 +1,0 @@
-const Case = require('../../entities/Case');
-
-/**
- * getCasesForRespondent
- * @param userId
- * @param applicationContext
- * @returns {*|Promise<*>}
- */
-exports.getCasesForRespondent = async ({ applicationContext }) => {
-  const user = applicationContext.getCurrentUser();
-
-  const cases = await applicationContext
-    .getPersistenceGateway()
-    .getCasesForRespondent({
-      userId: user.userId,
-      applicationContext,
-    });
-  return Case.validateRawCollection(cases);
-};
