+++ shared/src/persistence/dynamo/cases/saveCase.js
@@ -88,7 +88,7 @@ exports.saveCase = async ({ caseToSave, applicationContext }) => {
       applicationContext,
       pkId: caseToSave.respondent.respondentId,
       skId: caseToSave.caseId,
-      type: 'activeCase',
+      type: 'case',
     });
   }
